+++ src/games/planetwars/visualizer/src/renderers/expeditions.js
@@ -100,4 +100,5 @@ class ExpeditionRenderer {
   }
 }
 
+
 module.exports = ExpeditionRenderer;
